requirements.txt
autopep8==1.4.4
certifi==2022.12.7
chardet==3.0.4
cycler==0.10.0
idna==2.8
jsonpatch==1.24
jsonpointer==2.0
kiwisolver>=1.3.1
matplotlib==3.10.1
numpy>=1.23.5
Pillow>=7.1.0
pycodestyle==2.5.0
pyparsing==2.4.6
python-dateutil==2.8.1
pyzmq==26.3.0
scipy==1.8.0
six==1.13.0
torch>=2.0.0
torchfile==0.1.0
torchvision==0.21.0
tornado==6.0.3
tqdm==4.41.0
requests==2.32.3
urllib3==2.3.0
websocket-client==0.56.0